config -> config/v1
In preparation for config/v2, this is a straightforward refactor of config -> config/v1.

Tested by running apply on a few repos and go zero changes.
