+++ setup/setup.go
@@ -1,14 +1,14 @@
 package setup
 
 import (
-	"github.com/chanzuckerberg/fogg/config"
+	"github.com/chanzuckerberg/fogg/config/v1"
 	"github.com/chanzuckerberg/fogg/errs"
 	"github.com/chanzuckerberg/fogg/plugins"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/afero"
 )
 
-func Setup(fs afero.Fs, conf *config.Config) error {
+func Setup(fs afero.Fs, conf *v1.Config) error {
 	log.Debug("setting up plugins")
 	apply := func(name string, plugin *plugins.CustomPlugin) error {
 		log.Infof("Setting up plugin %s", name)
