+++ templates/templates.go
@@ -1,13 +1,13 @@
 package templates
 
 import (
-	"github.com/chanzuckerberg/fogg/config"
+	"github.com/chanzuckerberg/fogg/config/v1"
 	"github.com/gobuffalo/packr"
 )
 
 type T struct {
 	Account          packr.Box
-	Components       map[config.ComponentKind]packr.Box
+	Components       map[v1.ComponentKind]packr.Box
 	Env              packr.Box
 	Global           packr.Box
 	Module           packr.Box
@@ -18,9 +18,9 @@ type T struct {
 
 var Templates = &T{
 	Account: packr.NewBox("account"),
-	Components: map[config.ComponentKind]packr.Box{
-		config.ComponentKindTerraform:    packr.NewBox("component/terraform"),
-		config.ComponentKindHelmTemplate: packr.NewBox("component/helm_template"),
+	Components: map[v1.ComponentKind]packr.Box{
+		v1.ComponentKindTerraform:    packr.NewBox("component/terraform"),
+		v1.ComponentKindHelmTemplate: packr.NewBox("component/helm_template"),
 	},
 	Env:              packr.NewBox("env"),
 	Global:           packr.NewBox("global"),
